requirements.txt
langgraph>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
redis>=5.0.0
pydantic>=2.0.0
httpx>=0.27.0

